stacks/datamask-emr-launch/datalake/requirements.txt
aws-cdk-lib>=2.27.0
constructs>=10.1.0ß